Move discussion topic and discussion settings view methods.
